refactor: header links
