--- src/components/app-header/header-link/LinkLogo.jsx
@@ -1,11 +1,0 @@
-import { Link } from "react-router-dom";
-import { Logo } from "@ya.praktikum/react-developer-burger-ui-components";
-import { ROUTES } from "../../../utils";
-
-export const LinkLogo = () => {
-  return (
-    <Link to={ROUTES.ROOT} title="Stellar Burgers">
-      <Logo />
-    </Link>
-  );
-};
