+++ src/components/app-header/header-link/header-link.module.scss
@@ -1,7 +1,8 @@
 .link {
-  composes: text text_type_main-default text_color_inactive pt-4 pb-4 pr-5 from global;
+  composes: text text_type_main-default text_color_inactive pt-4 pb-4 pl-4 pr-4 from global;
   display: flex;
   flex-flow: row nowrap;
+  align-items: center;
 }
 
 .link::first-letter {
@@ -10,6 +11,8 @@
 
 .link_icon {
   composes: mr-2 from global;
+  display: flex;
+  align-items: center;
 }
 
 .active {
