+++ src/components/app-header/header-link/link-profile.jsx
@@ -3,8 +3,9 @@ import { ProfileIcon } from "@ya.praktikum/react-developer-burger-ui-components"
 import { HeaderLink } from "./header-link";
 import { ROUTES } from "../../../utils";
 import { selectName } from "../../../services/profile";
+import React from "react";
 
-export const LinkProfile = () => {
+export const LinkProfile = React.memo(() => {
   const name = useSelector(selectName);
   const profileLinkLabel = name || "Личный кабинет";
 
@@ -16,4 +17,4 @@ export const LinkProfile = () => {
       title="Личный кабинет"
     />
   );
-};
+});
